fixed uint warnings in tests, added rand feature attribute

+++ examples/checkerboard1d.rs
@@ -3,7 +3,7 @@ extern crate noisy;
 use noisy::gen::{ NoiseGen, Checkerboard };
 
 // Width of the output in columns.
-static WIDTH: int = 80;
+static WIDTH: usize = 80;
 
 // A five color gradient used for the output.
 static GRADIENT: [&'static str; 2] = [" ", "█"];
@@ -21,7 +21,7 @@ fn main() {
         val = (val + 1.0) * 0.5;
 
         // Print the columns in the row.
-        print!("{}", GRADIENT[val as uint]);
+        print!("{}", GRADIENT[val as usize]);
     }
     // Start a new row.
     println!("");
